requirements.txt
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
xarray==2024.3.0
netCDF4==1.6.5
rasterio==1.3.10
rio-cogeo==5.3.7
geopandas==0.14.4
pyproj==3.6.1
shapely==2.0.4
rioxarray==0.15.5
requests==2.32.3
python-dateutil==2.9.0.post0
typer==0.12.3
rich==13.7.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
pyyaml==6.0.2
tomli==2.0.1
azure-core==1.30.2
botocore==1.34.159

# Additional dependencies for enhanced pipeline
cdsapi==0.6.1
pyhdf==0.10.3
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2

